# web/openapi.json
{
 "components": {
  "schemas": {
   "AliasBody": {
    "additionalProperties": false,
    "properties": {
     "kind": {
      "enum": [
       "email",
       "name"
      ],
      "title": "Kind",
      "type": "string"
     },
     "value": {
      "maxLength": 320,
      "minLength": 1,
      "title": "Value",
      "type": "string"
     }
    },
    "required": [
     "kind",
     "value"
    ],
    "title": "AliasBody",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "MergePerson": {
    "additionalProperties": false,
    "properties": {
     "into_id": {
      "format": "uuid",
      "title": "Into Id",
      "type": "string"
     }
    },
    "required": [
     "into_id"
    ],
    "title": "MergePerson",
    "type": "object"
   },
   "PatchConversation": {
    "additionalProperties": false,
    "properties": {
     "priority_override": {
      "anyOf": [
       {
        "enum": [
         -1,
         0,
         1
        ],
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Priority Override"
     }
    },
    "required": [
     "priority_override"
    ],
    "title": "PatchConversation",
    "type": "object"
   },
...
   "PatchOrganization": {
    "additionalProperties": false,
    "properties": {
     "importance_user": {
      "anyOf": [
       {
        "maximum": 5.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Importance User"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     }
    },
    "title": "PatchOrganization",
    "type": "object"
   },
   "PatchPerson": {
    "additionalProperties": false,
    "properties": {
     "display_name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "importance_user": {
      "anyOf": [
       {
        "maximum": 5.0,
        "minimum": 1.0,
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Importance User"
     },
     "role_title": {
      "anyOf": [
       {
        "maxLength": 200,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Role Title"
     }
    },
    "title": "PatchPerson",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "Executive Context Assistant API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/conversations": {
   "get": {
    "operationId": "list_conversations_api_v1_conversations_get",
    "parameters": [
     {
      "in": "query",
      "name": "awaiting",
      "required": false,
      "schema": {
       "const": "user",
       "default": "user",
       "title": "Awaiting",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Conversations Api V1 Conversations Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Conversations"
   }
  },
  "/api/v1/conversations/{conversation_id}": {
   "get": {
    "operationId": "get_conversation_api_v1_conversations__conversation_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Conversation"
   },
   "patch": {
    "operationId": "patch_conversation_api_v1_conversations__conversation_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Conversation Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "If-Match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PatchConversation"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Conversation"
   }
  },
  "/api/v1/conversations/{conversation_id}/mark-handled": {
   "post": {
    "operationId": "mark_handled_api_v1_conversations__conversation_id__mark_handled_post",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Mark Handled"
   }
  },
...
  "/api/v1/meetings": {
   "get": {
    "operationId": "list_meetings_api_v1_meetings_get",
    "parameters": [
     {
      "in": "query",
      "name": "from",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "From"
      }
     },
     {
      "in": "query",
      "name": "to",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "To"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List Meetings Api V1 Meetings Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Meetings"
   }
  },
...
  "/api/v1/organizations/{org_id}": {
   "patch": {
    "operationId": "patch_organization_api_v1_organizations__org_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "org_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Org Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PatchOrganization"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "title": "Response Patch Organization Api V1 Organizations  Org Id  Patch"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Organization"
   }
  },
  "/api/v1/people": {
   "get": {
    "operationId": "list_people_api_v1_people_get",
    "parameters": [
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "in": "query",
      "name": "sort",
      "required": false,
      "schema": {
       "default": "importance",
       "enum": [
        "importance",
        "recent"
       ],
       "title": "Sort",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response List People Api V1 People Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List People"
   }
  },
  "/api/v1/people/{person_id}": {
   "get": {
    "description": "Person context: profile, identifiers and the open work involving them in both directions.",
    "operationId": "get_person_api_v1_people__person_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "person_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Person Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Person"
   },
   "patch": {
    "operationId": "patch_person_api_v1_people__person_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "person_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Person Id",
       "type": "string"
      }
     },
     {
      "in": "header",
      "name": "If-Match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PatchPerson"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Patch Person"
   }
  },
  "/api/v1/people/{person_id}/aliases": {
   "post": {
    "operationId": "add_alias_api_v1_people__person_id__aliases_post",
    "parameters": [
     {
      "in": "path",
      "name": "person_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Person Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AliasBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Add Alias"
   }
  },
  "/api/v1/people/{person_id}/merge": {
   "post": {
    "operationId": "merge_person_api_v1_people__person_id__merge_post",
    "parameters": [
     {
      "in": "path",
      "name": "person_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Person Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MergePerson"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Merge Person"
   }
  },
  "/api/v1/sources/{source_item_id}": {
   "get": {
    "description": "Source view within retention: metadata always, message text only while it is retained.",
    "operationId": "get_source_api_v1_sources__source_item_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "source_item_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Source Item Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Source Api V1 Sources  Source Item Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Source"
   }
  },
...
 }
}